style: refactor report_generator
