+++ deps/wazuh_testing/wazuh_testing/None
@@ -1,0 +1,1 @@
+{"metadata": {"n_agents": 1, "n_workers": 0}, "agents": {"criticals": [], "errors": [], "warnings": [], "wazuh-agentd": {"begin_status": {"connected": 1, "pending": 0, "disconnected": 0}, "end_status": {"connected": 1, "pending": 0, "disconnected": 0}, "ever_disconnected": 0, "ever_connected": 1, "ever_pending": 0, "mean_status_change_count": 0.0, "max_status_change_count": 0, "max_diff_ack_keep_alive": 0, "mean_diff_ack_keep_alive": 0.0, "metrics": {}, "begin": {}, "end": {}, "mean_msg_sent": 1628.0, "max_mean_msg_sent": 1628, "min_mean_msg_sent": 0, "min_msg_sent": 1625, "max_msg_sent": 1632, "mean_reg_cof_msg_sent": 0.0, "max_reg_cof_msg_sent": 0, "min_reg_cof_msg_sent": 0, "mean_msg_count": 1612.0, "max_mean_msg_count": 1612, "min_mean_msg_count": 0, "min_msg_count": 1612, "max_msg_count": 1612, "mean_reg_cof_msg_count": 0.0, "max_reg_cof_msg_count": 0, "min_reg_cof_msg_count": 0, "mean_msg_buffer": 0.0, "max_mean_msg_buffer": 0, "min_mean_msg_buffer": 0, "min_msg_buffer": 0, "max_msg_buffer": 0, "mean_reg_cof_msg_buffer": 0.0, "max_reg_cof_msg_buffer": 0, "min_reg_cof_msg_buffer": 0}}, "managers": {"criticals": [], "errors": [], "warnings": [{"manager0-iamd64-jenkins-util-aws-docker-instance_0": ["[ossec.log] 2022/02/17 16:06:30 wazuh-analysisd: WARNING: Mitre Technique ID 'T1089' not found in database."]}], "wazuh-remoted": {"mean_queue_size": 0.0, "max_mean_queue_size": 0, "min_mean_queue_size": 0, "min_queue_size": 0, "max_queue_size": 0, "mean_reg_cof_queue_size": 0.0, "max_reg_cof_queue_size": 0, "min_reg_cof_queue_size": 0, "mean_total_queue_size": 131072.0, "max_mean_total_queue_size": 131072, "min_mean_total_queue_size": 0, "min_total_queue_size": 131072, "max_total_queue_size": 131072, "mean_reg_cof_total_queue_size": 0.0, "max_reg_cof_total_queue_size": 0, "min_reg_cof_total_queue_size": 0, "mean_tcp_sessions": 1.0, "max_mean_tcp_sessions": 1, "min_mean_tcp_sessions": 0, "min_tcp_sessions": 1, "max_tcp_sessions": 1, "mean_reg_cof_tcp_sessions": 0.0, "max_reg_cof_tcp_sessions": 0, "min_reg_cof_tcp_sessions": 0, "mean_evt_count": 1612.0, "max_mean_evt_count": 1612, "min_mean_evt_count": 0, "min_evt_count": 1589, "max_evt_count": 1614, "mean_reg_cof_evt_count": 0.0, "max_reg_cof_evt_count": 0, "min_reg_cof_evt_count": 0, "mean_ctrl_msg_count": 18.0, "max_mean_ctrl_msg_count": 18, "min_mean_ctrl_msg_count": 0, "min_ctrl_msg_count": 15, "max_ctrl_msg_count": 21, "mean_reg_cof_ctrl_msg_count": 0.0, "max_reg_cof_ctrl_msg_count": 0, "min_reg_cof_ctrl_msg_count": 0, "mean_discarded_count": 0.0, "max_mean_discarded_count": 0, "min_mean_discarded_count": 0, "min_discarded_count": 0, "max_discarded_count": 0, "mean_reg_cof_discarded_count": 0.0, "max_reg_cof_discarded_count": 0, "min_reg_cof_discarded_count": 0, "mean_queued_msgs": 1031.0, "max_mean_queued_msgs": 1031, "min_mean_queued_msgs": 0, "min_queued_msgs": 1028, "max_queued_msgs": 1034, "mean_reg_cof_queued_msgs": 0.0, "max_reg_cof_queued_msgs": 0, "min_reg_cof_queued_msgs": 0, "mean_sent_bytes": 401589.0, "max_mean_sent_bytes": 401589, "min_mean_sent_bytes": 0, "min_sent_bytes": 401316, "max_sent_bytes": 401850, "mean_reg_cof_sent_bytes": 48.0, "max_reg_cof_sent_bytes": 48, "min_reg_cof_sent_bytes": 48, "mean_recv_bytes": 728727.0, "max_mean_recv_bytes": 728727, "min_mean_recv_bytes": 0, "min_recv_bytes": 718872, "max_recv_bytes": 730298, "mean_reg_cof_recv_bytes": 472.0, "max_reg_cof_recv_bytes": 472, "min_reg_cof_recv_bytes": 472, "mean_dequeued_after_close": 0.0, "max_mean_dequeued_after_close": 0, "min_mean_dequeued_after_close": 0, "min_dequeued_after_close": 0, "max_dequeued_after_close": 0, "mean_reg_cof_dequeued_after_close": 0.0, "max_reg_cof_dequeued_after_close": 0, "min_reg_cof_dequeued_after_close": 0, "keep_alives": {"ip-172-31-83-100.ec2.internal ": {"n_keep_alive": 19, "max_difference": 10, "mean_difference": 9.0, "last_keep_alive": "2022/02/17 16:09:21", "first_keep_alive": "2022/02/17 16:06:30", "remainder": 5}}}}}
