+++ deps/wazuh_testing/wazuh_testing/report.json
@@ -1,0 +1,158 @@
+{
+    "agents": {
+        "critical": [],
+        "error": [],
+        "warning": [],
+        "wazuh-agentd": {
+            "begin_status": {
+                "connected": 1,
+                "disconnected": 0,
+                "pending": 0
+            },
+            "end_status": {
+                "connected": 1,
+                "disconnected": 0,
+                "pending": 0
+            },
+            "ever_connected": 1,
+            "ever_disconnected": 0,
+            "ever_pending": 0,
+            "max_diff_ack_keep_alive": 0,
+            "max_mean_msg_buffer": 0,
+            "max_mean_msg_count": 1612,
+            "max_mean_msg_sent": 1628,
+            "max_msg_buffer": 0,
+            "max_msg_count": 1612,
+            "max_msg_sent": 1632,
+            "max_reg_cof_msg_buffer": 0,
+            "max_reg_cof_msg_count": 0,
+            "max_reg_cof_msg_sent": 0,
+            "max_status_change_count": 0,
+            "mean_diff_ack_keep_alive": 0.0,
+            "mean_msg_buffer": 0.0,
+            "mean_msg_count": 1612.0,
+            "mean_msg_sent": 1628.0,
+            "mean_reg_cof_msg_buffer": 0.0,
+            "mean_reg_cof_msg_count": 0.0,
+            "mean_reg_cof_msg_sent": 0.0,
+            "mean_status_change_count": 0.0,
+            "metrics": {},
+            "min_mean_msg_buffer": 0,
+            "min_mean_msg_count": 0,
+            "min_mean_msg_sent": 0,
+            "min_msg_buffer": 0,
+            "min_msg_count": 1612,
+            "min_msg_sent": 1625,
+            "min_reg_cof_msg_buffer": 0,
+            "min_reg_cof_msg_count": 0,
+            "min_reg_cof_msg_sent": 0
+        }
+    },
+    "managers": {
+        "critical": [],
+        "error": [],
+        "warning": [
+            {
+                "manager0-iamd64-jenkins-util-aws-docker-instance_0": [
+                    "[ossec.log] 2022/02/17 16:06:30 wazuh-analysisd: WARNING: Mitre Technique ID 'T1089' not found in database."
+                ]
+            }
+        ],
+        "wazuh-remoted": {
+            "keep_alives": {
+                "ip-172-31-83-100.ec2.internal ": {
+                    "first_keep_alive": "2022/02/17 16:06:30",
+                    "last_keep_alive": "2022/02/17 16:09:21",
+                    "max_difference": 10,
+                    "mean_difference": 9.0,
+                    "n_keep_alive": 19,
+                    "remainder": 5
+                }
+            },
+            "max_ctrl_msg_count": 21,
+            "max_dequeued_after_close": 0,
+            "max_discarded_count": 0,
+            "max_evt_count": 1614,
+            "max_mean_ctrl_msg_count": 18,
+            "max_mean_dequeued_after_close": 0,
+            "max_mean_discarded_count": 0,
+            "max_mean_evt_count": 1612,
+            "max_mean_queue_size": 0,
+            "max_mean_queued_msgs": 1031,
+            "max_mean_recv_bytes": 728727,
+            "max_mean_sent_bytes": 401589,
+            "max_mean_tcp_sessions": 1,
+            "max_mean_total_queue_size": 131072,
+            "max_queue_size": 0,
+            "max_queued_msgs": 1034,
+            "max_recv_bytes": 730298,
+            "max_reg_cof_ctrl_msg_count": 0,
+            "max_reg_cof_dequeued_after_close": 0,
+            "max_reg_cof_discarded_count": 0,
+            "max_reg_cof_evt_count": 0,
+            "max_reg_cof_queue_size": 0,
+            "max_reg_cof_queued_msgs": 0,
+            "max_reg_cof_recv_bytes": 472,
+            "max_reg_cof_sent_bytes": 48,
+            "max_reg_cof_tcp_sessions": 0,
+            "max_reg_cof_total_queue_size": 0,
+            "max_sent_bytes": 401850,
+            "max_tcp_sessions": 1,
+            "max_total_queue_size": 131072,
+            "mean_ctrl_msg_count": 18.0,
+            "mean_dequeued_after_close": 0.0,
+            "mean_discarded_count": 0.0,
+            "mean_evt_count": 1612.0,
+            "mean_queue_size": 0.0,
+            "mean_queued_msgs": 1031.0,
+            "mean_recv_bytes": 728727.0,
+            "mean_reg_cof_ctrl_msg_count": 0.0,
+            "mean_reg_cof_dequeued_after_close": 0.0,
+            "mean_reg_cof_discarded_count": 0.0,
+            "mean_reg_cof_evt_count": 0.0,
+            "mean_reg_cof_queue_size": 0.0,
+            "mean_reg_cof_queued_msgs": 0.0,
+            "mean_reg_cof_recv_bytes": 472.0,
+            "mean_reg_cof_sent_bytes": 48.0,
+            "mean_reg_cof_tcp_sessions": 0.0,
+            "mean_reg_cof_total_queue_size": 0.0,
+            "mean_sent_bytes": 401589.0,
+            "mean_tcp_sessions": 1.0,
+            "mean_total_queue_size": 131072.0,
+            "min_ctrl_msg_count": 15,
+            "min_dequeued_after_close": 0,
+            "min_discarded_count": 0,
+            "min_evt_count": 1589,
+            "min_mean_ctrl_msg_count": 0,
+            "min_mean_dequeued_after_close": 0,
+            "min_mean_discarded_count": 0,
+            "min_mean_evt_count": 0,
+            "min_mean_queue_size": 0,
+            "min_mean_queued_msgs": 0,
+            "min_mean_recv_bytes": 0,
+            "min_mean_sent_bytes": 0,
+            "min_mean_tcp_sessions": 0,
+            "min_mean_total_queue_size": 0,
+            "min_queue_size": 0,
+            "min_queued_msgs": 1028,
+            "min_recv_bytes": 718872,
+            "min_reg_cof_ctrl_msg_count": 0,
+            "min_reg_cof_dequeued_after_close": 0,
+            "min_reg_cof_discarded_count": 0,
+            "min_reg_cof_evt_count": 0,
+            "min_reg_cof_queue_size": 0,
+            "min_reg_cof_queued_msgs": 0,
+            "min_reg_cof_recv_bytes": 472,
+            "min_reg_cof_sent_bytes": 48,
+            "min_reg_cof_tcp_sessions": 0,
+            "min_reg_cof_total_queue_size": 0,
+            "min_sent_bytes": 401316,
+            "min_tcp_sessions": 1,
+            "min_total_queue_size": 131072
+        }
+    },
+    "metadata": {
+        "n_agents": 1,
+        "n_workers": 0
+    }
+}
